backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.6.0
sentence-transformers==3.0.1
openai==1.44.1
